requirements.txt
langchain
wikipedia
langchain_community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo_search